added state management to change from start menu to game board
