+++ src/Playbutton.jsx
@@ -1,17 +1,17 @@
-import { useState, useEffect } from "react";
-import "/Users/tpl822_6/project-techtonica/tictactoe_game2.0/src/App.css";
+// import { useState, useEffect } from "react";
+// // import "src/pages/App.css";
 
 
 
-export function Play(props){
+// export function Play(props){
    
-    return  <div className="play"><button><a href="#" target="_self" className="font-effect-neon">PLAY</a></button></div>
+//     return  <div className="play"><button><a href="#" target="_self" className="font-effect-neon">PLAY</a></button></div>
    
-}
+// }
 
-export function Name(){
-    return <div className="name" ><h1 className="font-effect-neon" >TIC TAC TOE</h1></div>
+// export function Name(){
+//     return <div className="name" ><h1 className="font-effect-neon" >TIC TAC TOE</h1></div>
 
-}
+// }
