+++ src/new_components/Cells.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Cells (props) {
+    
+
+    return <div className="cell" id={props.id} onClick={props.onClick}>{props.children}</div>
+}
+
+
+export default Cells;
+
