+++ src/pages/Home.jsx
@@ -1,18 +1,18 @@
-import { useState} from 'react'
-import { BrowserRouter, Routes, Route } from 'react-router-dom'
-import '/Users/tpl822_6/project-techtonica/tictactoe_game2.0/src/App.css'
-import "/Users/tpl822_6/project-techtonica/tictactoe_game2.0/src/compounds/Tictac.css"
-import Board from '../compounds/Board.jsx'
-import {Play, Name} from '/Users/tpl822_6/project-techtonica/tictactoe_game2.0/src/Playbutton.jsx'
+// import { useState} from 'react'
+// //import { BrowserRouter, Routes, Route } from 'react-router-dom'
+// //import '/Users/tpl822_6/project-techtonica/tictactoe_game2.0/src/App.css'
+// //import "/Users/tpl822_6/project-techtonica/tictactoe_game2.0/src/compounds/Tictac.css"
+// import Board from '../compounds/Board.jsx'
+// //import {Play, Name} from '/Users/tpl822_6/project-techtonica/tictactoe_game2.0/src/Playbutton.jsx'
 
 
-export function Home() {
-  const [count, setCount] = useState(0)
+// export function Home() {
+//   const [count, setCount] = useState(0)
 
-  return (
-     <div className="App">
-        <Board />
-      </div>
+//   return (
+//      <div className="App">
+//         <Board />
+//       </div>
 
-  )
-}
+//   )
+// }
